ai-predictor/requirements.txt
tensorflow==2.18.0
# tensorflow-gpu
torch
scikit-learn
pandas
numpy<2
psycopg2-binary
python-dotenv